finish admin-todos

+++ src/app/api/todos/route.ts
@@ -1,4 +1,5 @@
 import prisma from "@/app/lib/prisma";
+import { getUserSessionServer } from "@/auth/actions/auth-actions";
 import { NextResponse } from "next/server";
 import * as yup from "yup";
 
@@ -29,11 +30,16 @@ const postSchema = yup.object({
 });
 
 export async function POST(request: Request) {
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json("No autorizado", { status: 401 });
+  }
+
   try {
-    const body = await postSchema.validate(await request.json());
+    const { complete, description } = await postSchema.validate(await request.json());
 
     const todo = await prisma.todo.create({
-      data: body,
+      data: { complete, description, userId: user.id },
     });
 
     return NextResponse.json(todo);
@@ -43,8 +49,12 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSessionServer();
+  if (!user) {
+    return NextResponse.json("No autorizado", { status: 401 });
+  }
   try {
-    await prisma.todo.deleteMany({ where: { complete: true } });
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
 
     return NextResponse.json("Borrado exitoso");
   } catch (error) {
